+++ DOCS/CHANGELOG.md
@@ -77,6 +77,7 @@
   - log: `*`.
 - new optional dependencies: `fltk`, `flume`, `fontdue`, `gilrs`, `image`, `itertools`, `orion`, `ring`, `sdl2`, `sdl3`, `simdutf8`, `toml_edit`, `ureq`.
 - new profile: `wasm`.
+- new example: `js_web_api`.
 - add musl architectures to `check.rs` script.
 - add docs for monitored nightly features.